+++ crates/burn-remote/src/server/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod local_comm;
+pub(crate) mod pump;
 pub(crate) mod service;
 pub(crate) mod session;
 pub(crate) mod spawn;
@@ -7,7 +8,6 @@ pub(crate) mod time;
 pub(crate) mod transfer;
 pub(crate) mod worker;
 
-mod base;
 mod builder;
 #[cfg(feature = "iroh")]
 mod iroh;
